// frontend/src/app/generated/notification.ts
/**
 * RESTful API of iHRM
 * The RESTful API of the main core of the iHRM application. Add-ons will have their own API documentation.
 *
 * OpenAPI spec version: 2020.06.14
 * Contact: ihrm@imbus.tn
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Notification { 
    id: number;
    notifiedUser: number;
    description: string;
    userId?: number;
    url: string;
    notificationType: Notification.NotificationTypeEnum;
    createdAt: string;
    isRead: boolean;
}
export namespace Notification {
    export type NotificationTypeEnum = 'Leave' | 'Activity' | 'Project';
    export const NotificationTypeEnum = {
        Leave: 'Leave' as NotificationTypeEnum,
        Activity: 'Activity' as NotificationTypeEnum,
        Project: 'Project' as NotificationTypeEnum
    };
}